+++ src/main/java/com/acikgozKaan/LibraryAPI/dto/response/CategoryResponse.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public class CategoryResponse {
 
-    private int id;
+    private Long id;
     private String name;
     private String description;
     private List<Book> bookList;
@@ -15,18 +15,18 @@ public class CategoryResponse {
 
     }
 
-    public CategoryResponse(int id, String name, String description, List<Book> bookList) {
+    public CategoryResponse(Long id, String name, String description, List<Book> bookList) {
         this.id = id;
         this.name = name;
         this.description = description;
         this.bookList = bookList;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
